django-dashboard-atlantis-dark/api_app/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from platform_app.models import Project


class UserSerializer(serializers.ModelSerializer):

    class Meta:
        model = User
        fields = ('username', 'password')



class ProjectSerializer(serializers.ModelSerializer):

    class Meta:
        model = Project
        fields = '__all__'