django-dashboard-atlantis-dark/requirements.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.15
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
limesurveyrc2api==1.0.0
pdfminer3k==1.3.2
ply==3.11
pycodestyle==2.6.0
PyPDF2==1.26.0
python-decouple==3.3
pytz==2020.1
razorpay==1.2.0
requests==2.23.0
slate3k==0.5.3
sqlparse==0.3.1
Unipath==1.1
urllib3==1.25.9
whitenoise==5.0.1
